requirements.txt
numpy<1.19
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=1.0
torchvision
matplotlib
tensorflow
tensorboard
terminaltables
pillow
tqdm
